Reading simu and obs ts series works
The structure is not perfect yet. I put everything related to python into the
'python_libs' dir. The subfolder ('libsswim') has everything related to the SWIM model.
So there will be a subfolder for borg (libborg?) as well.

The test.py is a first shot but works. I want to have the test file on the level, that is
the project folder, because the real application will start here as well. However a
subfolder for test would be great. I think I have to set the PYTHONPATH to make that
working ... Lets see.