Use the json validator to check the configuration
All settings of the model are given via a json config file. The structure of this file is
of course nothing random but follows a schema definition. This definition is stored in
the package as a json file and follows json-schema as described here (http://json-schema.org/).

The python package 'jsonschema' is used to validate the defined schema with the user
given config file. That is really handy.