Time window function included
The window_ts function has now some tests, that include:

1. unit test for the function and their helper functions
2  helper functions that try to figure out where to set the start and end date for
truncation

Next: where do we put it in the config schema (json) and where is it called.

The window_ts function has now some tests, that include:

1. unit test for the function and their helper functions
2  helper functions that try to figure out where to set the start and end date for
truncation

Next: where do we put it in the config schema (json) and where is it called.