Add the option of an evaluation period
The configuration now eats some specifications for a evaluation period. That is the start
and the end date.
If the slot "evaluation_period" exists as a key with the objectives, the slots for start
and end dates are filled in the 'swim_objective' class.

We try to compare them with the start and end date and of the simulated time series and
cut the in a clever way.

A configuration with "no_error" is introduced.

Next: Issue a warnings if the objective function returns NAN values.