Parameter regions are working
The config has an option that specifies the number of Parameter regions. If nothing is
provided, it is set to __1__.

This option is considered when the limits of the parameters are read from the config and
at the moment the parameters are written to file for the next round. For BORG, this mean
nothing more than more parameters in total (factor parameter region).