Parameters work with the PIK version only.
The setup class creates a parameter template with a set of default
parameters. The list of parameters in the configuration json file must
have the same names as the parameter names in the template, otherwise we
stop.  All parameters are given only once and modifies for every
parameter region. That is something we should change in the future.
Writing the parameters is now a method of the setup class.