Usage of specific parameter regions only
We now specify a list of parameter regions we want to consider. Before,
all parameter regions were used. The list of ids is a required property
of the config schema. That enables us to calibrate only specific regions
out of many many more. By now that seems to work, but we need more tests
since that is quite a crucial feature.