Nested objectives (eg. Columns within a file)
In general, a single objective might be splittet into single ones
without specifying the input and output files over and over again.

If the slot 'nested' is specified, an array of strings where each string
represents a column name in the observed time series. The list of
objectives is than extended by each item of the array.

Currently, no checking is done whether the column exist in the observed
or simulated time series or not.

Check out the files *nested.json and their corresponding input files.